Remove endianness check and rewrite unpack_snps().
Since the data in snp_lookup[i].snp_array initialized with type uint8_t is simply copied in unpack_snps(), there is no need to consider endianness and the endianness check is removed.
In addition, unpack_snps() did not take data alignment into account, so it has been rewritten to do so.